+++ packages/liveblocks-react/src/factory.tsx
@@ -215,9 +215,9 @@ type RoomContext<
    * Storage.
    *
    * @example
-   * const [root] = useStorage();
+   * const root = useStorage();
    */
-  useStorage(): [root: LiveObject<TStorage> | null];
+  useStorage(): LiveObject<TStorage> | null;
 
   /**
    * useUpdateMyPresence is similar to useMyPresence but it only returns the function to update the current user presence.
@@ -443,7 +443,7 @@ export function createRoomContext<
     return room.getSelf();
   }
 
-  function useStorage(): [root: LiveObject<TStorage> | null] {
+  function useStorage(): LiveObject<TStorage> | null {
     const room = useRoom();
     const [root, setState] = React.useState<LiveObject<TStorage> | null>(null);
 
@@ -464,7 +464,7 @@ export function createRoomContext<
       };
     }, [room]);
 
-    return [root];
+    return root;
   }
 
   function useHistory(): History {
@@ -487,7 +487,7 @@ export function createRoomContext<
     key: TKey
   ): TStorage[TKey] | null {
     const room = useRoom();
-    const [root] = useStorage();
+    const root = useStorage();
     const rerender = useRerender();
 
     React.useEffect(() => {